requirements.txt
argparse
matplotlib
PyYAML
rich